feat: video, voice, and text chat functionality

+++ apps/web/app/lib/actions.ts
@@ -1,13 +1,17 @@
 "use server";
 
 import prisma from "@/app/lib/prisma";
+import { type RoomCode } from "@/app/lib/types";
+import axios from "axios";
 import { compareSync, hashSync } from "bcrypt-ts";
 import { revalidatePath } from "next/cache";
-import { createSession, updateSession, getSession } from "./session";
+import { createSession, getSession, updateSession } from "./session";
 import {
-  CreateReplyFields,
-  LoginFields,
   CourseFields,
+  CreateReplyFields,
+  HmsRoom,
+  LoginFields,
+  RoomData,
   RoomFields,
 } from "./types";
 
@@ -35,6 +39,10 @@ export async function authorizeUser(email: string, password: string) {
     throw new Error("Incorrect email or password");
   }
 
+  if (!user.password) {
+    throw new Error("Incorrect email or password");
+  }
+
   if (!compareSync(password, user.password)) {
     throw new Error("Incorrect email or password");
   }
@@ -56,7 +64,7 @@ export async function createCourse(data: CourseFields) {
       title,
     },
   });
-
+  revalidatePath("/platform/study-groups");
   return { success: true };
 }
 
@@ -143,9 +151,9 @@ export async function createChildReply(data: CreateReplyFields) {
 export async function deleteDiscussionReply(id: string) {
   const session = await getSession();
   if (!session) throw new Error("Unauthorized");
-
-  const dicussionId = (await prisma.reply.findUnique({ where: { id } }))
-    .discussionId;
+  const dicussionReply = await prisma.reply.findUnique({ where: { id } });
+  if (!dicussionReply) throw new Error("Reply not found");
+  const dicussionId = dicussionReply.discussionId;
   await prisma.reply.delete({ where: { id } });
   await prisma.discussion.update({
     where: { id: dicussionId },
@@ -212,6 +220,8 @@ export async function editDiscussion(
   revalidatePath("/platform/discussions");
 }
 
+// TODO: Persist the 100ms room data to the database
+/*
 export async function createRoom(data: RoomFields) {
   const session = await getSession();
   if (!session) throw new Error("Unauthorized");
@@ -220,7 +230,7 @@ export async function createRoom(data: RoomFields) {
   const room = await prisma.room.create({
     data: {
       name,
-      description,
+      description: description ?? "",
       course: {
         connect: {
           id: courseId,
@@ -242,6 +252,7 @@ export async function createRoom(data: RoomFields) {
 
   return room;
 }
+*/
 
 export async function deleteRoom(id: string, subject: string, code: number) {
   const session = await getSession();
@@ -286,4 +297,73 @@ export async function viewDiscussionPost(id: string) {
 
 export async function acceptCookies() {
   await updateSession({ cookiesAccepted: true });
+}
+
+export async function createHmsRoom(data: RoomData): Promise<HmsRoom> {
+  const session = await getSession();
+  if (!session) throw new Error("Unauthorized");
+
+  const { name, description } = data;
+  const token = process.env.HMS_MANAGEMENT_TOKEN;
+  const templateId = process.env.HMS_TEMPLATE_ID;
+
+  try {
+    const response = await axios.post(
+      `https://api.100ms.live/v2/rooms`,
+      {
+        name,
+        description,
+        template_id: templateId,
+      },
+      {
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
+      }
+    );
+
+    const roomId: string = response.data.id;
+    console.log(roomId);
+    const roomData = await axios.get(
+      `https://api.100ms.live/v2/rooms/${roomId}`,
+      {
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
+      }
+    );
+
+    const roomDetails = roomData.data;
+    const roomCode = await createRoomCode(roomDetails.id);
+    console.log(roomDetails);
+    return {
+      ...roomDetails,
+      codes: roomCode,
+    } as HmsRoom;
+  } catch (error) {
+    throw new Error("Failed to create room");
+  }
+}
+
+export async function createRoomCode(roomId: string): Promise<RoomCode[]> {
+  const session = await getSession();
+  if (!session) throw new Error("Unauthorized");
+  const token = process.env.HMS_MANAGEMENT_TOKEN;
+
+  try {
+    const response = await axios.post(
+      `https://api.100ms.live/v2/room-codes/room/${roomId}`,
+      {},
+      {
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
+      }
+    );
+    console.log(`Room ${roomId} code created`);
+    return response.data.data;
+  } catch (error) {
+    console.error(error);
+    throw new Error("Failed to create room code");
+  }
 }
